requirements.txt
pytorch==1.10.0
torchvision==0.11.0
torchaudio==0.10.0
cudatoolkit==11.3
numpy
Pillow==8.3.2
matplotlib
tqdm
tensorboardX
